scripts/pyproject.toml
[project]
name = "colab-enterprise-scripts"
version = "0.1.0"
description = "Utility scripts for Miro and project automation"
authors = [{ name = "Antigravity Agent" }]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
